requirements.txt
# --------- common --------- #
tqdm
pytest
biopython
numpy>=1.22.0

